Fixing build bug with no COW std::string / C++11
The ? operator in this line created a temporary copy of SysRoot that lived on StringRef's ctor stack. The c_str reference held by Prefix is invalid as soon as the ctor returns.

This only worked when run with a libstdc++ with a copy-on-write std:string implementation because the c_str pointer returned by Str.data() in StringRef ctor points to the same address as SysRoot.data(). With non-COW std::string, the data pointer held by Prefix pointed to the temporary string instance.

This has been already fixed on upstream llvm in the same way.