LHI and ORL pseudo assembler instructions
LHI (load high) loads the upper 18 bit of an immediate constant into a register using the R_XPIC_H18 relocation. After relocating, it behaves just like a normal load instruction with the modified immediate constant.

ORL (or low) bitwise or with the lower 14 bit of an immediate constant using the R_XPIC_L14 relocation. After relocating, it behaves just like a normal or instruction with the modified immediate constant.

With these two instructions (and a shift in between) it is possible to load abitrary 32 bit constants (most important: pointers) into a register.